Add a separate external view generation stage before sending out messages, so that the participants can use the external view when processing messages.
Also change the external view generation stage such that if a storage
node does not have state for a partition, in external view we will use
the default state in state model def as the place holder. This will make
sure that the external view always carries the information about the
final resourcekey locations, which is useful in the relay integration
scenario.